python/postgres_etl/requirements.txt
asyncpg~=0.29.0
python-dotenv~=1.0.0
pytest~=7.4.0
pytest-asyncio~=0.23.0
